[Bugfix] incorrect ammotype changing fix
When you try to reload weapon but count of ammos in your inventory is not enought for full magazine, ammo's iconcould instantly change to an icon of next ammotype.
Also the bug could lead to another kind of behavior. Take into your hands a weapon with not full mag and without ammo of current ammo type in the inventory. Start reload or changing ammotype and (without waiting until the action is finished) drop the weapon. Then take it back to the hands. The key for ammotype changing won't work.