Added option to create all sound sources on startup (#407)
* Added option to create all sound sources on startup
While researching issues #142 and #131, I used xperf to profile the stuttering that happens when a new AI comes online. During the stutter, most of the time is in CSoundRender_Core::i_create_source (and child calls).
This change adds a console variable that causes all sound sources to be created on startup, instead of being created when the sound is first needed. So it frontloads the work to make the actual play experience smoother. On my machine, on a release build, this adds 2.7 seconds to the startup time, but 90%+ of the stutter is eliminated when an AI comes online. These times will probably change depending on CPU and disk hardware.

* Renamed to snd_precache_all