Fixed some scriptable objects are not being registered on dedicated server (closes #517)
Normally, a number of objects (more particularly, scriptable objects)
are being registered in class_registrator.script, but on dedicated
server there's no scripts and engine registers these objects manually.

Actually, this is bad, because if you change something
class_registrator.script, you can forget to change the same in the
engine...

This is what happened to GSC, when they added new objects in scripts,
but forgot to change the engine.