xrCore/Memory/xr_memory_router.hpp: removed inline specifiers from new and delete overloads
Sorry. I did a big mistake.

https://en.cppreference.com/w/cpp/memory/new/operator_new
https://en.cppreference.com/w/cpp/memory/new/operator_delete
The behavior is undefined if more than one replacement is provided in
the program for any of the replaceable allocation function, or if a
replacement is defined with the inline specifier. The program is
ill-formed if a replacement is defined in namespace other than global
namespace, or if it is defined as a static non-member function at global
scope.