Fix CMake generator expressions comparing options
There are a few instances of CMake generator expressions that attempt to
check a configuration time option and convert it to a boolean, but use
the literal string name of the option instead of expanding it to the
value. Additionally, pass LUA_USE_POSIX through to the minilua
build-system since it is no longer on by default.