xrGame: fix ability to get a negative money balance after paid service
Why:
Money was checked when player press to a some question in dialog. If player by buy something at this moment, then your money state will be changed, but a condition for a service dialog wil not be rechecked.

How:
Just reset a talk dialog state to initial when player press to a trade button.

Issue: #1173