CMake: Fix determination of `ARCH_TYPE` variable
This was caused by using the `CMAKE_SIZEOF_VOID_P` variable before it
is actually set by cmake (after a call to project).

https://cmake.org/cmake/help/latest/variable/CMAKE_SIZEOF_VOID_P.html

> This is set to the size of a pointer on the target machine, and is
> determined when a compiled language is enabled. (...)

Also the `else(...)` construct in CMake works a bit different than the
original author expected since the expression is not evaluated and thus
we always set `ARCH_TYPE` to `x86` which is not correct.

Also took the liberty to properly output the `CMAKE_PROJECT_VERSION` and
the `ARCH_TYPE` variables.